Remove old HttpBody API This API was used between 4.0 and 4.5, but we can now get rid of it.
This patch should have no behavior changes, I only moved some methods and then removed the unused classes.
